Add GitHub Action for a contribution snake graph and update .gitignore
This commit introduces two new .gitignore files, one for the main directory and another inside the .idea directory, for improved organization and clearer rules. It features the addition of a GitHub Action in a .yml file that generates a contribution snake graph. This graphic visual aid helps to show the user's contribution activity in a fun and interactive way. README.md has also been significantly extended to include additional sections, icons, and styling for a more engaging and informative user profile.